% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{logliktest}
\alias{logliktest}
\title{Likelihood ratio test for two likfit models}
\usage{
logliktest(likfit1, likfit2)
}
\arguments{
\item{likfit1}{The first model returned from \link[geoR]{likfit}}

\item{likfit1}{The second model returned from \link[geoR]{likfit}.  It doesn't matter which is which.}
}
\value{
The p-value from the associated likelihood ratio test
}
\description{
Never have to remember which way to subtract again!
}
\note{
Don't even think about using non-nested models, or the wrath of Dr. Barry will haunt you for the rest of your days.
}
\examples{
data(lnLa)
model1 <- likfit(lnLa,cov.model="exponential",ini.cov.pars=c(1,18))
model2 <- likfit(lnLa,cov.model="exponential",ini.cov.pars=c(1,18),trend="1st")

logliktest(model1,model2)
logliktest(model2,model1)
}
\author{
Matt Tyers
}

