% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{makegrid}
\alias{makegrid}
\title{Makes a prediction grid from a geodata object}
\usage{
makegrid(geodata, length = 100, resolution = NULL)
}
\arguments{
\item{geodata}{The geodata object to use}

\item{length}{The number of points to use in the X- and Y- direction.  Defaults to 100, resulting in a 100 x 100 grid.}

\item{resolution}{Alternately, the grid can be defined by spacing with this dimension.  Specifying \code{resolution=10} would result in a grid in which points are spaced 10 units apart in the X- and Y- directions.  Defaults to \code{NULL}.}
}
\description{
Shaboom!
}
\examples{
data(lnLa)

grid <- makegrid(lnLa)
plot(lnLa$coords)
points(grid,pch='.')
}
\author{
Matt Tyers
}

