% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rasterblaster}
\alias{rasterblaster}
\title{Plots a kriging result in color!}
\usage{
rasterblaster(krigePlot, type = "predict", z = NULL, ramp = "rainbow",
  invert = F, alpha = 0.8, dark = 0.85, contour = T, xlab = "X coord",
  ylab = "Y coord", ...)
}
\arguments{
\item{krigePlot}{The kriging object to use, returned from \link[geoR]{krige.conv}}

\item{type}{The type of plot to produce.  Specifying \code{type="predict"} will produce a map of kriging predictions, specifying \code{type="se"} will produce a map of kriging standard errors, and specifying \code{type="lower95"} or \code{type="upper95"} will produce a map of lower or upper 95 percent CIs of kriging predictions.  Defaults to \code{"predict"}.}

\item{z}{Alternately, a vector of quantities can be used, with the same length as the prediction locations.  Defaults to \code{NULL}.}

\item{ramp}{The color ramp to use for plotting.  Currently, allowed ramps are \code{"grey"}, \code{"red"}, \code{"rainbow"}, \code{"stoplight"}, and \code{"heat"}.  Defaults to \code{"rainbow"}.}

\item{invert}{Whether to invert the color ramp.  Try it both ways, see if it looks better.}

\item{alpha}{Color transparency to use for plotting.  This ranges from 0 (completely transparent) to 1 (true color).  Defaults to 0.8.}

\item{dark}{Color saturation to use for plotting.  This ranges from 0 (black) to 1 (true color).  Defaults to 0.85.}

\item{contour}{Whether to add contour lines.  Defaults to \code{TRUE}.}

\item{...}{Additional plotting parameters}
}
\description{
Pass in a kriging result, and watch the world explode into technicolor, just like for Dorothy.
}
\examples{
data(krigePlot,locations)

rasterblaster(krigePlot)
rasterblaster(krigePlot,invert=T)
rasterblaster(krigePlot,alpha=.4)
rasterblaster(krigePlot,alpha=.4,dark=.7)
rasterblaster(krigePlot,ramp="grey")
}
\author{
Matt Tyers
}

